Move qsettings identifier stuff into a separate file (#42079)
