+++ qt/applications/workbench/workbench/identity.py
@@ -1,0 +1,13 @@
+# Mantid Repository : https://github.com/mantidproject/mantid
+#
+# Copyright &copy; 2026 ISIS Rutherford Appleton Laboratory UKRI,
+#   NScD Oak Ridge National Laboratory, European Spallation Source,
+#   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
+# SPDX - License - Identifier: GPL - 3.0 +
+#  This file is part of the mantid workbench.
+
+"""Side-effect-free application identity shared by Workbench processes."""
+
+ORGANIZATION = "mantidproject"
+ORG_DOMAIN = "mantidproject.org"
+APPNAME = "mantidworkbench"
